# Swagger.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"contact-controller","description":"Contact Controller"},{"name":"order-controller","description":"Order Controller"},{"name":"product-controller","description":"Product Controller"},{"name":"review-controller","description":"Review Controller"},{"name":"user-controller","description":"User Controller"}],"paths":{"/contact-requests":{"get":{"tags":["contact-controller"],"summary":"Gibt alle Kontaktanfragen zurück, ohne diese zu Filtern. Rückgabewert ist eine Liste als Json String formatiert.","operationId":"getAllContactRequestsUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["contact-controller"],"summary":"Fügt eine Kontaktanfrage hinzu. Wird die Operation erfolgreich ausgeführt, wird der Http-Statuscode 201 zurückgegeben.","operationId":"addContactUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"contact","description":"contact","required":true,"schema":{"$ref":"#/definitions/Contact"}}],"responses":{"201":{"description":"Created","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/contact-requests/bycustomernumber/{customerNumber}":{"get":{"tags":["contact-controller"],"summary":"Gibt eine Kontaktanfrage anhand der Kundennummer als Json String formatiert zurück.","operationId":"getSpecificContactByCustomerNumberUsingGET","produces":["application/json"],"parameters":[{"name":"customerNumber","in":"path","description":"customerNumber","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/contact-requests/{category}":{"get":{"tags":["contact-controller"],"summary":"Gibt eine Kontaktanfrage anhand der Kategorie als Json String formatiert zurück.","operationId":"getSpecificContactUsingGET","produces":["application/json"],"parameters":[{"name":"category","in":"path","description":"category","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/contact-requests/{customerNumber}":{"put":{"tags":["contact-controller"],"summary":"Ändert eine bestehende Kontaktabfrage ab. Wenn die Operation gelingt, wird der Http-Statuscode 200 zurückgegeben, sonst 400","operationId":"modifyContactUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"customerNumber","in":"path","description":"customerNumber","required":true,"type":"string"},{"in":"body","name":"newContact","description":"newContact","required":true,"schema":{"$ref":"#/definitions/Contact"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/contact-requests/{id}":{"delete":{"tags":["contact-controller"],"summary":"Löscht eine Kontaktanfrage anhand der gegebenen Id. Wenn die Operation gelingt, wird der Http-Statuscode 200 zurückgegeben.","operationId":"deleteContactUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingGET","produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"head":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingHEAD","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false},"post":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPOST","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPUT","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingDELETE","produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false},"options":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingOPTIONS","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false},"patch":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPATCH","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/orders":{"get":{"tags":["order-controller"],"summary":"Gibt alle Bestellungen zurück, ohne diese zu filtern. Rückgabewert ist eine Liste als Json String formatiert.","operationId":"getAllOrdersUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["order-controller"],"summary":"Fügt eine Bestellung hinzu. Wenn die Operation erfolgreich ist, wird der Http-Statuscode 200 zurückgegeben.","operationId":"addOrderUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"order","description":"order","required":true,"schema":{"$ref":"#/definitions/Orders"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/orders/{id}":{"put":{"tags":["order-controller"],"summary":"Ändert eine bestehende Bestellung ab. wenn die Operation erfolgreich ist, wird der Http-Statuscode 200 zurückgegeben, sonst 400.","operationId":"replaceOrderUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"},{"in":"body","name":"newOrder","description":"newOrder","required":true,"schema":{"$ref":"#/definitions/Orders"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["order-controller"],"summary":"Löscht eine Bestellung anhand der gegebenen Id. Ist die Operation erfolgreich, wird der Http-Statuscode 200 zurückgegeben.","operationId":"deleteOrderUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/orders/{product}":{"get":{"tags":["order-controller"],"summary":"Gibt eine Bestellung anhand eines Produktes zurück. Rückgabewert ist ein Bestellobjekt als Json String formatiert.","operationId":"getOrderByProductUsingGET","produces":["application/json"],"parameters":[{"name":"product","in":"path","description":"product","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/products":{"get":{"tags":["product-controller"],"summary":"Gibt alle Produkte aus dem Produkt Repository wieder, ohne diese zu filtern. Rückgabewert ist eine Liste als Json String formatiert.","operationId":"getAllProductsUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["product-controller"],"summary":"Fügt ein Produkt hinzu. Wenn die Operation erfolgreich war, wird der Http-Statuscode 200 zurückgegeben.","operationId":"addProductUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"product","description":"product","required":true,"schema":{"$ref":"#/definitions/Product"}}],"responses":{"201":{"description":"Created","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["product-controller"],"summary":"Ändert ein bestehendes Produkt ab. Wenn die Operation erfolgreich war, wird der Http-Statuscode 200 zurückgegeben, sonst 400.","operationId":"modifyProductUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"newProduct","description":"newProduct","required":true,"schema":{"$ref":"#/definitions/Product"}}],"responses":{"201":{"description":"Created","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/products/byname/{name}":{"get":{"tags":["product-controller"],"summary":"Gibt ein Produkt anhand seines Namens als Json String zurück.","operationId":"getSpecificProductByNameUsingGET","produces":["application/json"],"parameters":[{"name":"name","in":"path","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/products/{id}":{"get":{"tags":["product-controller"],"summary":"Gibt ein Produkt anhand der gegebenen Id zurück. Da die Id unique ist, wird hier nur ein Objekt als Json String zurückgegeben.","operationId":"getSpecificProductUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["product-controller"],"summary":"Löscht ein Produkt anhand der Id. Wenn die Operation erfolgreich ist, wird der Http-Statuscode 200 zurückgegeben.","operationId":"deleteProductUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/ratings/{rating}":{"get":{"tags":["review-controller"],"summary":"Gibt Reviews zurück, die dem gegebenen Rating entsprechen.","operationId":"getReviewsByRatingUsingGET","produces":["application/json"],"parameters":[{"name":"rating","in":"path","description":"rating","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/reviews":{"get":{"tags":["review-controller"],"summary":"Gibt alle Reviews aus dem Review Repository an, ohne diese zu filtern. Rückgabewert ist hier eine Liste als Json String formatiert.","operationId":"getAllReviewsUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["review-controller"],"summary":"Fügt ein Review hinzu. Wenn dies gelingt, wird der Http-Statuscode 201 zurückgegeben.","operationId":"addReviewUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"review","description":"review","required":true,"schema":{"$ref":"#/definitions/Review"}}],"responses":{"201":{"description":"Created","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/reviews/{id}":{"put":{"tags":["review-controller"],"summary":"Ersetzt ein existierendes Review. Wenn die Operation gelingt, wird der Http-Statuscode 200 zurückgegeben, sonst 400.","operationId":"changeReviewUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"},{"in":"body","name":"newReview","description":"newReview","required":true,"schema":{"$ref":"#/definitions/Review"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["review-controller"],"summary":"Löscht ein Review anhand der gegebenen Id. Wenn dies gelingt, wird der Http-Statuscode 200 zurückgegeben.","operationId":"deleteReviewUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/reviews/{username}":{"get":{"tags":["review-controller"],"summary":"Gibt alle REviews zurück, bei dem der Username dem des übergebenen entspricht.","operationId":"getReviewsByIdUsingGET","produces":["application/json"],"parameters":[{"name":"username","in":"path","description":"username","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/users":{"get":{"tags":["user-controller"],"summary":"Gibt alle Nutzer aus dem User Repository an, ohne diese zu filtern. Rückgabewert ist hier eine Liste als Json String formatiert.","operationId":"getAllUsersUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["user-controller"],"summary":"Fügt einen Nutzer hinzu. Rückgabewert ist hier beim erfolgreichen anlegen der Http-Statuscode 201.","operationId":"addUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"user","description":"user","required":true,"schema":{"$ref":"#/definitions/User"}}],"responses":{"201":{"description":"Created","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["user-controller"],"summary":"Überschreibt einen Nutzer. Sollte der Vorgang erfolgreich sein, wird der Http-Statuscode 201 zurückgegeben, sonst 400.","operationId":"modifyUserUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"newUser","description":"newUser","required":true,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/users/byname/{name}":{"get":{"tags":["user-controller"],"summary":"Gibt einen Nutzer anhand seines Namens zurück. Zurtst wird der Vorname überprüft, danach der Nachname. Rückgabewert ist ein Json formatiertes Objekt.","operationId":"getSpecificUserByNameUsingGET","produces":["application/json"],"parameters":[{"name":"name","in":"path","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/users/{id}":{"get":{"tags":["user-controller"],"summary":"Gibt einen Nutzer anhand seiner Id zurück. Da die Id unique ist, gibt es immer nur einen Nutzer als Json formatiertes Objekt zurück.","operationId":"getSpecificUserUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["user-controller"],"summary":"Löscht einen Nutzer anhand seiner Id. Wenn das Löschen erfolgreich ist, wird der Http-Statuscode 200 zurückgegeben.","operationId":"deleteUserUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}}},"definitions":{"Contact":{"type":"object","properties":{"category":{"type":"string"},"customerNumber":{"type":"string"},"description":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"id":{"type":"integer","format":"int32"},"lastName":{"type":"string"},"subject":{"type":"string"}},"title":"Contact"},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}},"title":"ModelAndView"},"Orders":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"product":{"type":"string"},"purchaser":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"totalCosts":{"type":"integer","format":"int32"}},"title":"Orders"},"Product":{"type":"object","properties":{"brand":{"type":"string"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"price":{"type":"integer","format":"int32"},"since":{"type":"integer","format":"int32"},"url":{"type":"string"}},"title":"Product"},"Review":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"rating":{"type":"integer","format":"int32"},"text":{"type":"string"},"username":{"type":"string"}},"title":"Review"},"User":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"id":{"type":"integer","format":"int32"},"lastName":{"type":"string"},"password":{"type":"string"},"userName":{"type":"string"}},"title":"User"},"View":{"type":"object","properties":{"contentType":{"type":"string"}},"title":"View"}}}